🔧(ci) run tox in parallel
This ask tox to run in parallel on three vCPU.

The pros using tox are the very convenient way to configure testing
for both local tests and CI, the cons are it makes CI tests quite slow.

I may propose another pull request for CI optimization later.

The use of tox --parallel reduces time at low cost since the Django
tests adds lot of testing environments so it is good enough for now.