🧑‍💻(django) allow standalone django
This adds the possibility for developpers to run the Django integration
test project in standalone mode.

This is not a feature of the `social_edu_federation` library but it
may be useful for developers to see how things are linked and how
to use the library or customize it or make pull requests.