🧑‍💻(django) allow fake idp port override
This allows to replace the port used to generate the metadata for the fake
IDP to provide an easy way to use it inside a docker container.

If the Django server runs on port 8000 and the docker maps 8000 to 11000
then we expect to use port 11000 ouside docker and therefore to have
port 11000 in the metadata.